requirements.txt
Flask==3.0.3
email-validator==2.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
WTForms==3.1.2
bcrypt==4.1.3
cryptography==42.0.8
python-dotenv==1.0.1
pytest==8.2.2
pytest-flask==1.3.0
bandit==1.7.9
safety==3.2.3
